Move test_histogram to its own module
