+++ external/vcm/tests/test_calc.py
@@ -14,7 +14,6 @@ from vcm.calc.thermo import (
 )
 from vcm.calc.calc import local_time, apparent_source
 from vcm.cubedsphere.constants import COORD_Z_CENTER, COORD_Z_OUTER
-from vcm.calc.histogram import histogram
 
 
 @pytest.mark.parametrize("toa_pressure", [0, 5])
@@ -122,18 +121,6 @@ def test_apparent_source():
         s_dim="forecast_time",
     )
     assert Q1_forecast3 == pytest.approx((2.0 / (15 * 60)) - (4.0 / 60))
-
-
-def test_histogram():
-    data = xr.DataArray(
-        np.reshape(np.arange(0, 40, 2), (5, 4)), dims=["x", "y"], name="temperature"
-    )
-    coords = {"temperature_bins": [0, 30]}
-    expected_count = xr.DataArray([15, 5], coords=coords, dims="temperature_bins")
-    expected_width = xr.DataArray([30, 10], coords=coords, dims="temperature_bins")
-    count, width = histogram(data, bins=[0, 30, 40])
-    xr.testing.assert_equal(count, expected_count)
-    xr.testing.assert_equal(width, expected_width)
 
 
 def test_mass_streamfunction():
